+++ DSCResources/SqlServer/SqlServer.schema.psm1
@@ -96,6 +96,9 @@ Configuration SqlServer
     Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion '2.4.0.0'
     . "$resourcePath\Windows.SecurityOption.ps1"
 
+    Import-DscResource -ModuleName AccessControlDsc -ModuleVersion 1.4.0.0
+    . "$resourcePath\Windows.AccessControl.ps1"
+
     Import-DscResource -ModuleName PSDscResources -ModuleVersion 2.10.0.0
     . "$resourcePath\windows.Script.skip.ps1"
 }